refactor(pdf-parser): modularize VLM pipeline and simplify conversion logic
Extract VLM-related conversion logic into a new `VlmConversionPipeline` class for improved modularity and maintainability. Update `PDFConverter` to use the pipeline, removing the inline `convertWithVlm` method. Refactor and consolidate tests for VLM handling to align with the new structure.
